requirements.txt
fastapi
sqlmodel
dotenv
maxapi
aiogram
requests
sqlalchemy
sqlmodel
yookassa
alembic
email-validator
aiohttp_socks
aiohttp
argon2-cffi